requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
groq
google-generativeai==0.8.2
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.11
wikipedia==1.4.0
PyPDF2==3.0.1
Pillow
numpy
lxml==5.1.0
pytz==2024.1
